backend/requirements.txt
# Core FastAPI and web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0

# AWS and Bedrock
boto3>=1.34.0
botocore>=1.34.0

# LangChain for tools
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0


pydantic>=2.5.0
pydantic-settings>=2.1.0

loguru>=0.7.2

httpx>=0.25.0

pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
pydantic-settings>=2.1.0
pymupdf
langsmith>=0.1.0
langchain-core>=0.1.0